管理画面　コントローラーのコメント追加

+++ app/Http/Controllers/Admin/Admin/List/DeleteController.php
@@ -5,14 +5,28 @@ namespace App\Http\Controllers\Admin\Admin\List;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\Admin;
+use Illuminate\Http\RedirectResponse;
 
 class DeleteController extends Controller
 {
     /**
-     * Handle the incoming request.
+     * 管理者削除
+     *
+     * @param Request $request
+     * @return RedirectResponse
      */
-    public function __invoke(Request $request)
+    public function __invoke(Request $request): RedirectResponse
     {
+        // ルートユーザーは削除不可
+        if ($request->route('id') == 1) {
+            return redirect()
+                ->route('admin.admin.list')
+                ->with([
+                    'severity' => 'warning',
+                    'message' => 'この管理者は削除出来ません',
+                ]);
+        }
+
         $admin = Admin::Id($request->route('id'))->first();
         $name = $admin->name;
         $admin->delete();
